Bump 0.12.x to 0.13.0
Feat: Add derive-getters and update config

The config is updated to include `Option` types for `users` and
`current_user`. The `derive-getters` crate is added to provide getter
methods for accessing these fields. Additionally, the README and Config
documentation files have been updated to reflect the changes made to
the config file.